backend/pyproject.toml
[project]
name = "backend-python"
version = "0.1.12"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.40.0",
    "websockets>=13.0,<15.0",
    "firebase-admin>=6.6.0",
    "google-genai>=1.0.0",
    "python-multipart>=0.0.20",
]

[dependency-groups]
dev = [
    "ruff>=0.9.3",
]

[tool.ruff]
line-length = 88
target-version = "py313"
